Fixes IPv6 address encoding.
I noticed this when trying to encode the IP address `2a01:4f9:c011:87ad:0:0:0:0` in `ocaml-dns`, but decoding it resulted in `ffff:ffff:c011:b100::1`.

As `b` is a signed 32 bit integer, if b is negative `Int64.of_int32 b` will create a negative 64-bit integer.

That is, if we have the IP address `0:0:8000:0:0:0:0:0`, with `b=1000000000000000 0000000000000000`, then `Int64.of_int32 b` will equal `1111111111111111 1111111111111111 1000000000000000 0000000000000000` instead of `0000000000000000 0000000000000000 1000000000000000 0000000000000000`.

The same is true of `Int64.of_int32 d`.

This isn't a problem for `a` and `b` as they are shifted left 32 bits anyway.

This commit fixes this bug.

See https://github.com/mirage/ocaml-ipaddr/issues/113.